use wide layout on exchange data export page
